backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
openai>=1.3.5
python-multipart>=0.0.6
pydantic>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.24.1
pydub>=0.25.1
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.4.0
aiosqlite>=0.19.0
email-validator>=2.0.0
certifi>=2023.7.22
pyjwt>=2.8.0
cryptography>=41.0.0
